# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummaryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Lista documentos del proyecto para OWNER o EDITOR",
        "tags": [
          "uml-documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "El canonicalModel no cumple el schema 0.1.0."
          },
          "409": {
            "description": "El nombre ya existe dentro del proyecto."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Crea y valida un documento UML canónico",
        "tags": [
          "uml-documents"
        ]
      }
    },
    "/projects/{projectId}/documents/{documentId}": {
      "get": {
        "operationId": "DocumentsController_get",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Documento inexistente o perteneciente a otro proyecto."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Consulta el modelo canónico dentro del proyecto autorizado",
        "tags": [
          "uml-documents"
        ]
      },
      "put": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO o canonicalModel inválidos."
          },
          "404": {
            "description": "Documento inexistente o perteneciente a otro proyecto."
          },
          "409": {
            "description": "Conflicto de revisión o de nombre único.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/RevisionConflictResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/HttpConflictResponseDto"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Actualiza por compare-and-swap e incrementa la revisión una vez",
        "tags": [
          "uml-documents"
        ]
      },
      "delete": {
        "operationId": "DocumentsController_delete",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Documento inexistente o perteneciente a otro proyecto."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Elimina un documento como OWNER o EDITOR; revisiones caen en cascada",
        "tags": [
          "uml-documents"
        ]
      }
    },
    "/projects/{projectId}/documents/{documentId}/xmi/import/preview": {
      "post": {
        "operationId": "DocumentsController_previewXmiImport",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/XmiPreviewUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmiImportPreviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "XMI malformado, inválido o semánticamente incompatible."
          },
          "401": {
            "description": "Access token o sesión inválidos."
          },
          "403": {
            "description": "El actor no tiene permiso para importar el documento."
          },
          "404": {
            "description": "Documento inexistente o perteneciente a otro proyecto."
          },
          "408": {
            "description": "El procesamiento XMI excedió el presupuesto permitido."
          },
          "413": {
            "description": "El archivo XMI supera los límites permitidos."
          },
          "415": {
            "description": "El archivo no declara un tipo XML/XMI aceptado."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Previsualiza un XMI sin persistir cambios",
        "tags": [
          "uml-documents"
        ]
      }
    },
    "/projects/{projectId}/documents/{documentId}/xmi/import/apply": {
      "post": {
        "operationId": "DocumentsController_applyXmiImport",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ApplyXmiImportUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "XMI, hash o modelo canónico inválido, o advertencias del preview sin confirmar (acknowledgeWarnings)."
          },
          "401": {
            "description": "Access token o sesión inválidos."
          },
          "403": {
            "description": "El actor no tiene permiso para importar el documento."
          },
          "404": {
            "description": "Documento inexistente o perteneciente a otro proyecto."
          },
          "408": {
            "description": "El procesamiento XMI excedió el presupuesto permitido."
          },
          "409": {
            "description": "Conflicto de revisión, nombre o lock activo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionConflictResponseDto"
                }
              }
            }
          },
          "413": {
            "description": "El archivo XMI supera los límites permitidos."
          },
          "415": {
            "description": "El archivo no declara un tipo XML/XMI aceptado."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Aplica un XMI validado mediante compare-and-swap",
        "tags": [
          "uml-documents"
        ]
      }
    },
    "/projects/{projectId}/documents/{documentId}/xmi/export": {
      "get": {
        "operationId": "DocumentsController_exportXmi",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expectedRevision",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/xmi+xml": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "XMI generado desde el snapshot autorizado y validado.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "example": "private, no-store",
                  "type": "string"
                }
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "X-Content-Type-Options": {
                "schema": {
                  "example": "nosniff",
                  "type": "string"
                }
              },
              "X-Document-Revision": {
                "schema": {
                  "type": "integer"
                }
              },
              "X-XMI-Profile-Version": {
                "schema": {
                  "type": "string"
                }
              },
              "X-XMI-SHA256": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Revisión solicitada o snapshot persistido inválido."
          },
          "401": {
            "description": "Access token o sesión inválidos."
          },
          "403": {
            "description": "El actor no tiene permiso para exportar el documento."
          },
          "404": {
            "description": "Documento inexistente o perteneciente a otro proyecto."
          },
          "409": {
            "description": "La revisión esperada no coincide con la revisión persistida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionConflictResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Exporta el snapshot UML persistido como XMI UTF-8",
        "tags": [
          "uml-documents"
        ]
      }
    },
    "/projects/{projectId}/documents/{documentId}/exports/spring-boot": {
      "post": {
        "operationId": "DocumentsController_exportSpringBoot",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportSpringBootDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ZIP Spring Boot generado desde el snapshot autorizado.",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "private, no-store"
                }
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string",
                  "example": "nosniff"
                }
              },
              "X-Document-Revision": {
                "schema": {
                  "type": "integer"
                }
              },
              "X-Generator-Version": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Snapshot persistido u opciones de exportacion invalidos."
          },
          "401": {
            "description": "Access token o sesión inválidos."
          },
          "403": {
            "description": "El actor no tiene permiso para exportar el documento."
          },
          "404": {
            "description": "Proyecto o documento inexistente, eliminado o inaccesible."
          },
          "408": {
            "description": "La generación Spring Boot excedió el tiempo máximo."
          },
          "409": {
            "description": "La revision esperada no coincide con la revision persistida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionConflictResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Se alcanzo el limite local de frecuencia o concurrencia."
          },
          "500": {
            "description": "No fue posible generar el ZIP solicitado."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Exporta el snapshot UML persistido como ZIP Spring Boot determinista",
        "tags": [
          "uml-documents"
        ]
      }
    }
  },
  "info": {
    "title": "UML Class Designer API",
    "description": "API interna de la herramienta CASE; no es el backend Spring Boot generado.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refresh-cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uml_refresh"
      }
    },
    "schemas": {
...
      "RevisionConflictResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Document revision conflict."
          },
          "currentRevision": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "required": [
          "message",
          "currentRevision"
        ]
      },
...
      "DocumentSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "string",
            "example": "0.1.0"
          },
          "revision": {
            "type": "number",
            "minimum": 0
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "updatedById": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "schemaVersion",
          "revision",
          "createdById",
          "updatedById",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160,
            "example": "Modelo de dominio"
          },
          "canonicalModel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CanonicalUmlModel"
              }
            ],
            "description": "Documento UML canónico validado contra el JSON Schema 0.1.0."
          }
        },
        "required": [
          "name",
          "canonicalModel"
        ]
      },
...
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "expectedRevision": {
            "type": "number",
            "minimum": 0,
            "description": "Revisión persistida que el cliente leyó."
          },
          "canonicalModel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CanonicalUmlModel"
              }
            ],
            "description": "Documento UML canónico completo, nunca un snapshot de React Flow."
          }
        },
        "required": [
          "expectedRevision",
          "canonicalModel"
        ]
      },
...
    }
  }
}